A task in a workflow can now have volumes.
The intention here is to specify all the volumes at task level, that are common for all
actions within that task. The volumes at task level will be mounted in all actions. However,
if a user has the same mount at action level, it will get the higher preference and will
override the one at task level.