Drop method: rebase in favor of the default (merge) strategy
Rebase will rebase all the commits on top of the base branch, but we
don't want that. Having the merge commit is a nice indicator in the
history of groups of commits that were part of a PR. By having the
`strict: smart+fastpath` setting we are already guaranteeing that the PR
has been updated against the base branch.

See:
https://doc.mergify.io/actions.html#merge
and
https://doc.mergify.io/merge-action.html#strict-merge