docker-compose: remove EFK
Fetching EFK adds several minutes to the setup time and start-times.

I think most hackers will want to just see the logs on the console,
and be interested in EFK once things are more prod-like. Using EFK
means the user actually has to do some initial config on their
cluster, and it doesn't just work out of the box.

One problem is OSIE wants to talk specifically to ElasticSearch, but
I wonder if we could swap it out with syslog and a syslog listener?