Disable tk client setup  if an --help flag is present
An up and running Tinkerbell Server should not be a requirement for the
`help` command to work.

This is a bit tricky because our client is globally instantiated as part
of the root command.

I made it to work for:

```
$ ./cmd/tink-cli/tink-cli workflow --help
```

Previously the command returned `2020/06/08 11:16:25 undefined
TINKERBELL_CERT_URL`. Now it returns the help command.

I think this should work the same way, but it is hard to do, so I am
skipping it for now.
```
$ ./cmd/tink-cli/tink-cli workflow
```

Signed-off-by: Gianluca Arbezzano <gianarb92@gmail.com>