bump timeout from 5min to 10min
There is a timeout for the Vagrant Setup guide test at 5minutes. If the
expected set of statuses does not show up in 5 minutes the tests gets
marked as failed.

It works locally but in CI it fails, so this is an attempt to see if it
is only a resource issue

Signed-off-by: Gianluca Arbezzano <gianarb92@gmail.com>