Add reflection to grpc service
I discovered this useful introspection technique trying to use grpcurl

https://github.com/fullstorydev/grpcurl

I know one of the use case of the HTTP API is because they are easy to
use and more familiar.

I think we just need to write and share a bit more about how to use the
gRPC. The gRPC community is pretty big and there tools that can help
with that.

Signed-off-by: Gianluca Arbezzano <gianarb92@gmail.com>
Signed-off-by: Jonathan Wong <jonawong@equinix.com>