change primary key from 'created_at' to 'id'
We can have multiple events generated at a given time, therefore the value of
'created_at' field will be same for each event. This violates the unique constraint
on "events_pkey", when these events are add to the events table.

The migration changes the primary key on events table from 'created_at' to `id`,
which will always be unique for each event generated at any point in time.

Signed-off-by: Gaurav Gahlot <gauravgahlot0107@gmail.com>