Use MOQ to generate grpc mock for hardware proto
I was looking for something easy to use to generate moq structure from
an interface.

I discovered a great project from the awesome Mat called Moq
https://github.com/matryer/moq.

It generated structures starting from an interface following the
strategy we use already for the db package.

This commit contains a test case that prove how to use that mock for:

  tink hardware get

Signed-off-by: Gianluca Arbezzano <gianarb92@gmail.com>