Remove FullClient from command New function
I thought it was a good idea to pass the FullClient to all the commands
that interact with Tinkerbell server as part of the `New*` function.

But it requires to attempt a connection to the tink-server very early
and commands that do not need a client such as "docs", "help" were still
requiring a running tink-server, and we do not want it.

Signed-off-by: Gianluca Arbezzano <gianarb92@gmail.com>