db/migration: Add test for basic Migration struct validation
We've had previous commits that introduced migrations where the func
name and migration id did not match, that could be pretty confusing.
Same thing can happen with the file name too. It just seems like a good
idea to make all the timestamp/names match so we can easily go from
migrations in the db to func to file.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>