Remove deploy directory
As discussed on Slack and briefly in #418 we are not sure we need the
deploy directory now that we have sandbox.

The initial idea about leaving it in the tink repository was to provide
an environment that developers can use to contribute to Tinkerbell.

I am not sure if it evolved in that way and it is kind of making
maintenance hard because it diverges from what we have in sandbox.

Overall the proposal is to remove it from tink.

Signed-off-by: Gianluca Arbezzano <gianarb92@gmail.com>