Remove gRPC TLS cert
The way we to TLS in Tinkerbell is just a complication.
It does not increase security because everybody that knows how to
download the certificate from the tink-server can connect.

It can't be used to identify the requester and it means that
is not a carrier that we can use for authorisation moving forward.

With a couple of other maintainers we would like to remove TLS
completely for now until we can figure out a better design for it.

Signed-off-by: Gianluca Arbezzano <gianarb92@gmail.com>