Add simplified docker-compose:
The previous deploy directory was spinning up
more than just tink. This is a duplicate of what is
in the sandbox repo. This commit adds a basic docker-compose
for just tink server and tink cli. Easy to use make target
`run` will build the binaries first and then up the stack.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>