do not attach the action container stdout/stderr to the current process
when tink-worker is configured with --capture-action-logs=false the action
container stdout/stderr must not be attached to the current process.

that configuration is used when there is an existing docker log-driver, which
will capture all the containers logs without tink-worker help.

this prevents the action container logs from being duplicated in the
tink-worker container logs and in the action container docker logs.

Signed-off-by: Rui Lopes <rgl@ruilopes.com>