Use tools.go based method to install controller-gen
No need to have 2 different ways to manage go based tooling, we already have
tools.go based install using simple `go install` calls.

This also has the benefit that make has proper knowledge of the dependency on
the binary and how to generate it so we don't need to do recursive make which
is harmful.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>