Update .envrc with better stuff
Make use of direnv's stdlib having `has` which replaces the old `which ... &>/dev/null`.
This also sources env vars from .env files which should not be added to git and allow for
local user to override/add things without git trying to track it.
Finally, add ./bin to both PATH and GOBIN to avoid contamination of global GOBIN.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>