tink-worker: Simplify connection to tink-server
Previously we were calling client.Setup() (which returns if GetConnection()
errors) and later followed by calling client.GetConnection in a loop until a
successful connection is made. This made no sense since if the first call to
client.Setup fails we return from RunE. So if the first call succeeds in
creating tinkerbell connection, why do we want to throw it away and try again?
Makes no sense, so I replaced with a simple call to GetConnection

Signed-off-by: Manuel Mendez <mmendez@equinix.com>