make: Better test target
Why do we clean the test cache? It should not be necessary (I haven't found it be
necessary).

Also get rid of -v as its just noisy and never useful. When there is not test failure
I don't really care to see *all* the tests being run. Yet when there *is* a test failure
we potentially have to scroll over all the passing tests just to get to the output of the
failed one. Why fight go here? Go will expand the test output for the failed tests anyway
so its just doesn't make any sense to run with -v.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>