Send tink-worker logs to stdout instead of stderr:
In HookOS, normally, all Docker logs are sent via syslog
to Smee. When a container writes its logs to stderr the
Docker syslog driver will set the severity to "ERR". This
then shows up in Smee's logs. User are consistently confused
as to what this severity means and more often than not, they
believe the log line to be an error. By writing tink-worker
logs to stdout the severity in syslog will be "INFO".

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>